api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
google-cloud-bigquery==3.1.0
pydantic==2.11.4
pydantic-settings==2.9.1
python-dotenv==1.1.0